# Store: smaller block size for snappy decoding