@MichaHoffmann 2023-06-21
> Tests failure seems related.

fixed and adapted benchmark results!