@MichaHoffmann 2023-06-27
Better left to the experts: https://github.com/thanos-io/thanos/pull/6475/files